Refactors map-request-params per TGM
