--- config/middleware/chatbot/required-params.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const configVars = {};
-
-configVars.containerProperty = 'body';
-configVars.requiredParams = [
-  'phone',
-];
-module.exports = configVars;
