+++ test/lib/middleware/map-request-params.test.js
@@ -8,13 +8,13 @@ const sinon = require('sinon');
 const sinonChai = require('sinon-chai');
 const httpMocks = require('node-mocks-http');
 const logger = require('winston');
-const Promise = require('bluebird');
 const rewire = require('rewire');
 
 const stubs = require('../../utils/stubs');
-const stathat = require('../../../lib/stathat');
-const helpers = require('../../../lib/helpers');
-const config = require('../../../config/middleware/chatbot/map-request-params');
+const config = require('../../../config/middleware/receive-message/map-request-params');
+
+const userId = stubs.getUserId();
+const campaignId = stubs.getCampaignId();
 
 // setup "x.should.y" assertion style
 chai.should();
@@ -26,17 +26,10 @@ const mapRequestParams = rewire('../../../lib/middleware/map-request-params');
 // sinon sandbox object
 const sandbox = sinon.sandbox.create();
 
-// stubs
-const stathatStub = () => true;
-const getKeywordStub = Promise.resolve(stubs.getJSONstub('fetchKeyword'));
-const getBroadcastStub = Promise.resolve(stubs.getJSONstub('fetchBroadcast'));
 
 // Setup!
 test.beforeEach((t) => {
   stubs.stubLogger(sandbox, logger);
-  sandbox.stub(stathat, 'postStat').returns(stathatStub);
-  sandbox.stub(helpers, 'getKeyword').returns(getKeywordStub);
-  sandbox.stub(helpers, 'getBroadcast').returns(getBroadcastStub);
 
   // setup req, res mocks
   t.context.req = httpMocks.createRequest();
@@ -50,32 +43,23 @@ test.afterEach((t) => {
   t.context = {};
 });
 
-test('mapRequestParams should populate the contentful broadcast object and campaignId when it gets a broadcast_id', async (t) => {
+test('mapRequestParams should populate the req object with variables defined in config', async (t) => {
   // setup
   const next = sinon.stub();
   const middleware = mapRequestParams(config);
-  const broadcastObject = stubs.getJSONstub('fetchBroadcast');
-  const campaign = broadcastObject.fields.campaign.fields;
-  t.context.req[config.containerProperty].broadcast_id = stubs.getBroadcastId();
+  const paramsMap = config.paramsMap;
+  const bodyParams = Object.keys(paramsMap);
+  t.context.req.body = {
+    userId,
+    campaignId,
+  };
 
   // test
   await middleware(t.context.req, t.context.res, next);
-  t.context.req.contentfulObjects.broadcast.should.be.equal(broadcastObject);
-  t.context.req.campaignId.should.be.equal(campaign.campaignId);
-  next.should.have.been.called;
-});
-
-test('mapRequestParams should populate the contentful keyword object and campaignId from the keyword if no broadcast_id is found', async (t) => {
-  // setup
-  const next = sinon.stub();
-  const middleware = mapRequestParams(config);
-  const keywordObject = stubs.getJSONstub('fetchKeyword');
-  const campaign = keywordObject.fields.campaign.fields;
-  t.context.req[config.containerProperty].keyword = stubs.getKeyword();
-
-  // test
-  await middleware(t.context.req, t.context.res, next);
-  t.context.req.contentfulObjects.keyword.should.be.equal(keywordObject);
-  t.context.req.campaignId.should.be.equal(campaign.campaignId);
+  bodyParams.forEach((bodyParamName) => {
+    t.context.req.should.have.property(paramsMap[bodyParamName]);
+  });
+  t.context.req.userId.should.be.equal(userId);
+  t.context.req.campaignId.should.be.equal(campaignId);
   next.should.have.been.called;
 });
