+++ test/lib/middleware/required-params.test.js
@@ -8,7 +8,7 @@ const sinon = require('sinon');
 const sinonChai = require('sinon-chai');
 const httpMocks = require('node-mocks-http');
 
-const config = require('../../../config/middleware/chatbot/required-params');
+const config = require('../../../config/middleware/receive-message/required-params');
 const helpers = require('../../../lib/helpers');
 
 // setup "x.should.y" assertion style
